\name{sparktex-package}
\alias{sparktex-package}
\title{Create LaTeX sparklines from R}
\description{Generate sparktex output for LaTeX from input value}
\details{
\tabular{ll}{
Package: \tab sparktex\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-06-06\cr
License: \tab GPL-2\cr
}
}
\author{
Thomas J. Leeper, with contributions from Andy Barbour

Maintainer: Thomas J. Leeper <thosjleeper@gmail.com>
}